+++ resources/views/datatable.blade.php
@@ -9,14 +9,14 @@
         <x-livewire-tables::tools>
             <x-livewire-tables::tools.sorting-pills />
             <x-livewire-tables::tools.filter-pills />
-            <x-livewire-tables::tools.toolbar />
+            <x-livewire-tables::tools.toolbar :$filterGenericData />
         </x-livewire-tables::tools>
 
         <x-livewire-tables::table>
             <x-slot name="thead">
                 <x-livewire-tables::table.th.reorder x-cloak x-show="currentlyReorderingStatus" />
                 <x-livewire-tables::table.th.bulk-actions :displayMinimisedOnReorder="true" />
-                <x-livewire-tables::table.th.row-contents />
+                <x-livewire-tables::table.th.collapsed-columns />
 
                 @foreach($columns as $index => $column)
                     @continue($column->isHidden())
@@ -28,7 +28,7 @@
             </x-slot>
 
             @if($this->secondaryHeaderIsEnabled() && $this->hasColumnsWithSecondaryHeader())
-                <x-livewire-tables::table.tr.secondary-header :rows="$rows" />
+                <x-livewire-tables::table.tr.secondary-header :rows="$rows" :$filterGenericData />
             @endif
             @if($this->hasDisplayLoadingPlaceholder())
                 <x-livewire-tables::includes.loading colCount="{{ $this->columns->count()+1 }}" />
@@ -41,7 +41,7 @@
                 <x-livewire-tables::table.tr wire:key="{{ $tableName }}-row-wrap-{{ $row->{$this->getPrimaryKey()} }}" :row="$row" :rowIndex="$rowIndex">
                     <x-livewire-tables::table.td.reorder x-cloak x-show="currentlyReorderingStatus" wire:key="{{ $tableName }}-row-reorder-{{ $row->{$this->getPrimaryKey()} }}" :rowID="$tableName.'-'.$row->{$this->getPrimaryKey()}" :rowIndex="$rowIndex" />
                     <x-livewire-tables::table.td.bulk-actions wire:key="{{ $tableName }}-row-bulk-act-{{ $row->{$this->getPrimaryKey()} }}" :row="$row" :rowIndex="$rowIndex"/>
-                    <x-livewire-tables::table.td.row-contents wire:key="{{ $tableName }}-row-collapsed-{{ $row->{$this->getPrimaryKey()} }}" :rowIndex="$rowIndex" />
+                    <x-livewire-tables::table.td.collapsed-columns wire:key="{{ $tableName }}-row-collapsed-{{ $row->{$this->getPrimaryKey()} }}" :rowIndex="$rowIndex" />
 
                     @foreach($columns as $colIndex => $column)
                         @continue($column->isHidden())
@@ -54,7 +54,7 @@
                     @endforeach
                 </x-livewire-tables::table.tr>
 
-                <x-livewire-tables::table.row-contents :row="$row" :rowIndex="$rowIndex" />
+                <x-livewire-tables::table.collapsed-columns :row="$row" :rowIndex="$rowIndex" />
             @empty
                 <x-livewire-tables::table.empty />
             @endforelse
@@ -62,9 +62,9 @@
             @if ($this->footerIsEnabled() && $this->hasColumnsWithFooter())
                 <x-slot name="tfoot">
                     @if ($this->useHeaderAsFooterIsEnabled())
-                        <x-livewire-tables::table.tr.secondary-header :rows="$rows" />
+                        <x-livewire-tables::table.tr.secondary-header :rows="$rows" :$filterGenericData />
                     @else
-                        <x-livewire-tables::table.tr.footer :rows="$rows" />
+                        <x-livewire-tables::table.tr.footer :rows="$rows"  :$filterGenericData />
                     @endif
                 </x-slot>
             @endif