+++ src/Views/Traits/Helpers/FilterHelpers.php
@@ -255,4 +255,19 @@ trait FilterHelpers
     {
         return $tableName.'-filter-'.$filterType.'-'.$this->getKey().($extraData != '' ? '-'.$extraData : '').($this->hasCustomPosition() ? '-'.$this->getCustomPosition() : '');
     }
+
+    public function getGenericDisplayData(): array
+    {
+        return $this->genericDisplayData;
+    }
+
+    public function getFilterDisplayData(): array
+    {
+        return array_merge($this->getGenericDisplayData(), ['filter' => $this]);
+    }
+
+    public function getViewPath(): string
+    {
+        return $this->viewPath;
+    }
 }