+++ tests/Views/Filters/DateFilterTest.php
@@ -146,4 +146,14 @@ final class DateFilterTest extends FilterTestCase
 
         $this->assertSame('2023-03-01', self::$filterInstance->getFilterDefaultValue());
     }
+
+    /**
+     * @test
+     */
+    public function can_set_custom_filter_view(): void
+    {
+        $this->assertSame('livewire-tables::components.tools.filters.date', self::$filterInstance->getViewPath());
+        self::$filterInstance->setCustomView('test-custom-filter-view');
+        $this->assertSame('test-custom-filter-view', self::$filterInstance->getViewPath());
+    }
 }